+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 2.0.0
+- Breaking: Removing Protractor and phantomjs-prebuilt as dependencies  
+(Should be installed separately if needed)
+
 ## 1.0.1
 - Randomized report name (for parallel execution)
 - Aggregate more than on test in a unique HTML report
 
 ## 1.0.0
-Initial public release.
+- Initial public release.